requirements.txt
streamlit
langchain
langchain-openai
langchain-experimental
sqlalchemy
pandas
python-dotenv
pyodbc
plotly